+++ private/modules/dispatch.py
@@ -22,11 +22,15 @@ def show_help(is_error=0):
     print get_help_text()
     
 def get_help_text():
+  from helper_tasks import list_members
+  
   text = '\n'.join(['Usage:',
-    '  $ private module_path [flags] <command route> [args]\n',
-    'Flags',
+    '  $ private module_path [flag] <command route> [args]',
+    '\nFlags',
     ' -h    show help.',
-    ' -p    execute a command with partial args.'
-  ])
+    ' -p    execute a command with partial args.',
+    '\nMembers\n',
+  ]
+  ) + '  ' + '\n  '.join([('%s  %s' % (name, desc))[:60] for name, desc in list_members('')])
   
   return text