+++ scripts/examples/advanced/wrapping.py
@@ -1,0 +1,9 @@
+"""
+An example fro wrapping private based modules
+"""
+
+from private import interface
+
+import simple # the privated script
+
+interface.call('task1 arg2=1', True)